codegen: find structs file by patterns
When code generating, we ought to ensure that ent/pro structs don't get
processed for codec code generation.

This uses a basic file pattern match rather than having a static list of
non-ent/pro files that we may forget to update.

The trade-offs are:
* Missed file causes streaming RPC to go through non-optimized path
* including an ent/pro in code generation accidentally causes a
  `go generate` failure, because references to ent/pro tagged files.

As such, it makes more sense to be greedy in match as we would detect
any accidental ent/pro leakage as part of code generate.

This ports https://github.com/hashicorp/nomad/pull/5646 to Enterprise
repo.