logmon: Refactor fifo access for windows safety
On unix platforms, it is safe to re-open fifo's for reading after the
first creation if the file is already a fifo, however this is not
possible on windows where this triggers a permissions error on the
socket path.

In this commit we add an explicit path for attempting to connect to
fifos before creating them. This makes fifo connection possible on
windows, and simplifies the error path for fifo connection.