agent: fix error message on -dev=connect
The stdlib `flag` package doesn't allow non-bool flags to be passed an
empty value. The `-dev=connect` flag was working around this by
setting `IsBoolVar` on the `Value` implementation, but this resulted
in unclear error messages about an "invalid boolean". To avoid this,
we'll manipulate the args directly to set `-dev=true` if the `-dev`
flag is present before we parse.