Update json version in lockfile
In #6113, I had to change the version of this gem because
I was getting this error, which is happening again now
in the Netlify deployment:

Your bundle is locked to json (1.8.3.1), but that version could not be found in
any of the sources listed in your Gemfile. If you haven't changed sources, that
means the author of json (1.8.3.1) has removed it. You'll need to update your
bundle to a version other than json (1.8.3.1) that hasn't been removed in order
to install.