Move kind property to task
I put this property in the wrong place. I’ve been unable to
properly update the tests that verify the presence of the
proxy tag, sadly. The changes I make to models in Mirage
aren’t being reflected in the mock API responses. So I’ve
skipped the relevant tests for now.