Change test to make every task state a proxy
It appears the problem is that despite my attempts to force
the first task in the list to be the one with the tag so
it can be asserted against, it’s not always the first. So
this should make every task be of kind proxy, so asserting
against the first will work.