vendor: update go-dockerclient to 1.4.1
The go-dockerclient library that was used did not support to disable
swapping for a docker-container.
When MemorySwappiness was set to 0, the value was not passed to the
docker-client. Therefore the swappiness setting of the parent cgroup was
used.

Version 1.4.1 contains the change https://github.com/fsouza/go-dockerclient/pull/776.
MemorySwappiness is changed to a pointer type. If MemorySwappiness is
null, the value will not be passed. If it's 0, swappiness is set to 0.

This will allow to disable swapping for docker containers.