docker: disallow docker-containers to use swap
According to the nomad documentation docker containers cannot use swap.
This was not the case. Container were swapping.

The swappiness value for docker containers was unset. This had the
effect that the swappiness value is inherited from the parent cgroup,
where swappiness is enabled by default.

Setting memory.memsw.limit_in_bytes (MemorySwap) to the same value then
memory.limit_in_bytes (task.Resources.LinuxResources.MemoryLimitBytes)
is not sufficient to disable swapping of a container.
If the host system runs out of memory, the container will start swapping
to make more RAM available for other processes if swappiness is not set
to 0.