nomad exec: Allow customizing environment variables
When execing into a task, it's useful to be able to set some special
environment variables, namely `TERM` and language environment variable.

In this PR, we expose `env` parameter that takes an array of `KEY=VALUE`
values that gets thread all the way through to executors exec handling.

The main purpose is to allow configuring `TERM` for purposes of web ui,
as the appropriate value would depend on the terminal library and not
necessarily user friendly.

CLI
----------

This PR doesn't expose this configuration to CLI due to some
complexities:

* It's not as valuable to set environment variables there
(i.e. users can export their environment variables before running
commands)
* The `-e` flag is already used for escape character, and I'm unsure if
how to best proceed; I can see the appeal of repurposing `-e` to
env-variables as it's the most common thing
* We use the standard library `flags` package that doesn't support
string array flag to allow for multiple environment variables nicely.

As such, I've punted on supporting the CLI for another time as the need
arises.

External drivers and compatibility
----------

This PR makes some changes to the exec endpoint in drivers.  These
changes are binary backward compatible: existing drivers continue to run
against latest nomad binary as the grpc handler ignores the new `env`
field.  The drivers continue to compile as normal against the version
they vendored or depended on.

Though, when plugin authors upgrade the dependency they will need to
update code to either handle the `env` field or explicitly ignore it.