Remove target for iframe base element
The target attribute was causing table-sorting to break
on Netlify as it was attempting to navigate.

The base target change caused links with href="#" and target="_self"
to start navigating within the iframe to a nested copy of the
whole Storybook interface. Removing the target did the same,
and changing to target="_parent" caused the parent page to
return to the first story. So this moves to using an inert
Javascript href instead.

Links that should actually navigate now need target="_parent" to
ensure navigation happens on the parent page.